+++ changes/ocd-4588.sql
@@ -1,3 +1,0 @@
-UPDATE openchpl.certification_criterion_attribute
-SET privacy_security_framework = false
-WHERE criterion_id = 210;